feat: toggle purchase balance and total balance int to double

+++ src/main/java/com/oocl/easymovie/entity/PurchasePoint.java
@@ -25,8 +25,8 @@ public class PurchasePoint {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private Long userId;
-    private Integer balance;
-    private Integer historyTotal;
+    private Double balance;
+    private Double historyTotal;
 
     @CreationTimestamp
     private Date createTime;
